methods/spider-agent-snow/requirements.txt
httpx==0.27.2
docker
flaml
transformers
openai
gymnasium
rich
openai==1.31.0
langchain
chromadb
tiktoken
diskcache
pypdf
unstructured
fuzzywuzzy
python-Levenshtein
playwright
sentence-transformers
faiss-cpu
minify_html
pandas
llama-cpp-python
tenacity
accelerate
optimum
auto-gptq
rpyc
mysql-connector-python
google-cloud-bigquery
duckdb
psycopg2
psycopg2-binary
google.generativeai
plotly
dashscope
groq
textgrad